Update templates with hooks, layout, and inline errors
This updates the templates generated by superglue_rails to use the usePage
hooks. The values are still avail through the page components props.

We also update the input components to use inline errors out-of-the-box, and added
a simple layout component to use a shared flash.
