+++ superglue_rails/lib/generators/superglue/install/install_generator.rb
@@ -25,8 +25,10 @@ module Superglue
         say "Copying application_visit.js file to #{app_js_path}"
         copy_file "#{__dir__}/templates/js/application_visit.js", "#{app_js_path}/application_visit.js"
 
-        say "Copying components.js file to #{app_js_path}"
-        copy_file "#{__dir__}/templates/js/components.js", "#{app_js_path}/components.js"
+        say "Copying components to #{app_js_path}"
+        copy_file "#{__dir__}/templates/js/inputs.js", "#{app_js_path}/components/Inputs.js"
+        copy_file "#{__dir__}/templates/js/layout.js", "#{app_js_path}/components/Layout.js"
+        copy_file "#{__dir__}/templates/js/components.js", "#{app_js_path}/components/index.js"
 
         say "Copying jsconfig.json file to #{app_js_path}"
         copy_file "#{__dir__}/templates/js/jsconfig.json", "jsconfig.json"
