+++ superglue_rails/lib/generators/superglue/install/templates/js/layout.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import { useSelector } from 'react-redux'
+
+export const Layout = ({children}) => {
+  const flash = useSelector((state) => state.flash)
+
+  return (
+    <div>
+      {flash.success && <p>{flash.success}</p>}
+      {flash.notice && <p>{flash.notice}</p>}
+      {flash.error && <p>{flash.error}</p>}
+
+      {children}
+    </div>
+  )
+}
