+++ superglue_rails/lib/generators/superglue/view_collection/templates/js/edit.js
@@ -1,20 +1,30 @@
 import React from 'react'
-import { Form, TextField } from '@javascript/components'
-
-// import { useSelector } from 'react-redux'
+import {
+  Form,
+  Layout,
+  <%- attributes.each do |attr| -%>
+  <%= js_component(attr)%>,
+  <%- end -%>
+} from '@javascript/components'
+import { useSelector } from 'react-redux'
+import { usePage } from '@thoughtbot/superglue'
 
 export default function <%= js_plural_table_name(:upper) %>Edit ({
   // visit,
   // remote,
-  form,
-  errors,
-  <%= js_singular_table_name %>Path,
-  <%= js_plural_table_name %>Path,
 }) {
+  const {
+    form,
+    <%= js_singular_table_name %>Path,
+    <%= js_plural_table_name %>Path,
+  } = usePage().data
+
   const { inputs, props, extras } = form
+  const validationErrors = useSelector((state) => state.flash["<%= js_singular_table_name%>FormErrors"])
+
   return (
-    <div>
-      <Form {...props} extras={extras} data-sg-visit>
+    <Layout>
+      <Form {...props} extras={extras} validationErrors={validationErrors} data-sg-visit>
         <%- attributes.each do |attr| -%>
         <<%= js_component(attr)%> {...inputs.<%= attr.column_name.camelize(:lower)%>} label="<%= attr.column_name.humanize %>" errorKey="<%= attr.column_name %>" />
         <%- end -%>
@@ -23,6 +33,6 @@ export default function <%= js_plural_table_name(:upper) %>Edit ({
 
       <a href={<%= js_singular_table_name %>Path} data-sg-visit>Show</a>
       <a href={<%= js_plural_table_name %>Path}  data-sg-visit>Back</a>
-    </div>
+    </Layout>
   )
 }
