+++ superglue_rails/lib/generators/superglue/view_collection/templates/js/new.js
@@ -1,19 +1,29 @@
 import React from 'react'
-import { Form, TextField } from '@javascript/components'
-// import { useSelector } from 'react-redux'
+import { 
+  Form, 
+  Layout,
+  <%- attributes.each do |attr| -%>
+  <%= js_component(attr)%>,
+  <%- end -%>
+} from '@javascript/components'
+import { useSelector } from 'react-redux'
+import { usePage } from '@thoughtbot/superglue'
 
 export default function <%= js_plural_table_name(:upper) %>New({
   // visit,
   // remote
-  form,
-  errors,
-  <%= js_plural_table_name %>Path,
 }) {
+  const {
+    form,
+    <%= js_plural_table_name %>Path
+  } = usePage().data
+
   const { inputs, props, extras } = form
+  const validationErrors = useSelector((state) => state.flash["<%= js_singular_table_name%>FormErrors"])
 
   return (
-    <div>
-      <Form {...form.props} data-sg-visit>
+    <Layout>
+      <Form {...props} extras={extras} validationErrors={validationErrors} data-sg-visit>
         <%- attributes.each do |attr| -%>
         <<%= js_component(attr)%> {...inputs.<%= attr.column_name.camelize(:lower)%>} label="<%= attr.column_name.humanize %>" errorKey="<%= attr.column_name %>" />
         <%- end -%>
@@ -21,6 +31,6 @@ export default function <%= js_plural_table_name(:upper) %>New({
       </Form>
 
       <a href={<%= js_plural_table_name %>Path} data-sg-visit>Back</a>
-    </div>
+    </Layout>
   )
 }
