+++ superglue_rails/lib/generators/superglue/view_collection/templates/js/show.js
@@ -1,20 +1,21 @@
 import React from 'react'
-import { useSelector } from 'react-redux'
+import { Layout } from '@javascript/components'
+import { usePage } from '@thoughtbot/superglue'
 
 export default function <%= js_plural_table_name(:upper) %>Show({
   // visit,
   // remote,
+}) {
+  const {
   <%- attributes_list_with_timestamps.select{|attr| attr != :id }.each do |attr| -%>
   <%=attr.camelize(:lower)%>,
   <%- end -%>
   edit<%= js_singular_table_name(:upper) %>Path,
   <%= js_plural_table_name %>Path
-}) {
-  const flash = useSelector((state) => state.flash)
+  } = usePage().data
 
   return (
-    <div>
-      <p id="notice">{flash && flash.notice}</p>
+    <Layout>
       <%- attributes_list_with_timestamps.select{|attr| attr != :id }.each do |attr| -%>
       <p>
         <strong><%= attr.capitalize %>:</strong>
@@ -23,6 +24,6 @@ export default function <%= js_plural_table_name(:upper) %>Show({
       <%- end -%>
       <a href={ edit<%= js_singular_table_name(:upper) %>Path } data-sg-visit>Edit</a>
       <a href={ <%= js_plural_table_name %>Path } data-sg-visit>Back</a>
-    </div>
+    </Layout>
   )
 }
